Optimized code
removed methods from VC to separate file, remade searchPass func logic if there is no pass to hack
