+++ iOS6-HW15-Alexander.Strelkov/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
