+++ iOS6-HW15-Alexander.Strelkov/Resources/Assets.xcassets/Contents.json
