+++ iOS6-HW15-Alexander.Strelkov/Resources/Base.lproj/LaunchScreen.storyboard
