+++ iOS6-HW15-Alexander.Strelkov/Resources/Info.plist
