+++ iOS6-HW15-Alexander.Strelkov/Sources/Scenes/AppDelegate.swift
@@ -1,0 +1,15 @@
+//
+//  AppDelegate.swift
+//  iOS6-HW15-Alexander.Strelkov
+//
+//  Created by Alexandr Strelkov on 11.07.2022.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+}
+
+
