+++ iOS6-HW15-Alexander.Strelkov/Sources/Scenes/SceneDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  SceneDelegate.swift
+//  iOS6-HW15-Alexander.Strelkov
+//
+//  Created by Alexandr Strelkov on 11.07.2022.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+}
+
